backend/requirements.txt
openai>=1.6.1
langchain>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.2.post1
python-dotenv==1.0.0
chromadb>=0.4.22
pytest==7.4.4
numpy
pandas
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.7
markdown>=3.5.0
